[ENHANCEMENT] Consolidate requiresVSCode tests into a single combined run
- Extracts shared git helpers into gitTestUtils.ts, removing duplicate implementations across four test files.

- Adds vscodeRegression.test.ts which calls all five requiresVSCode test suites sequentially in one VS Code extension-host launch,
  - reduces sequential VS Code invocations from 5 to 1.

- Updates registry to replace the five individual requiresVSCode entries with a single regression_vscode entry.
