+++ src/tests/registry.ts
@@ -150,13 +150,6 @@ export const TEST_GROUPS: TestGroup[] = [
                 testModule: './incomingNonConflictRegression.test.js',
             },
             {
-                id: 'regression_logic_metadata_renumber',
-                description: 'Non-conflict metadata + renumber execution_count correctness',
-                notebooks: ['demo_base.ipynb', 'demo_current.ipynb', 'demo_incoming.ipynb'],
-                testModule: './logicRegression.test.js',
-                requiresVSCode: true,
-            },
-            {
                 id: 'regression_null_current_cell_indexing',
                 description: 'Cell indexing when current is null (delete/modify conflict)',
                 notebooks: ['demo_base.ipynb', 'demo_current.ipynb', 'demo_incoming.ipynb'],
@@ -169,31 +162,10 @@ export const TEST_GROUPS: TestGroup[] = [
                 testModule: './settingsMatrix.test.js',
             },
             {
-                id: 'regression_status_indicators',
-                description: 'Status bar + file decoration update on startup/conflict/add',
+                id: 'regression_vscode',
+                description: 'VS Code regressions: status bar, file decorations, git index statuses, DU/UD/AU/UA pick flows, merge logic',
                 notebooks: ['02_base.ipynb', '02_current.ipynb', '02_incoming.ipynb'],
-                testModule: './statusIndicatorsRegression.test.js',
-                requiresVSCode: true,
-            },
-            {
-                id: 'regression_unmerged_status_matrix',
-                description: 'Detect all unmerged statuses and enforce DD discoverability exclusion',
-                notebooks: ['02_base.ipynb', '02_current.ipynb', '02_incoming.ipynb'],
-                testModule: './unmergedStatusMatrix.test.js',
-                requiresVSCode: true,
-            },
-            {
-                id: 'regression_du_ud_pick_one',
-                description: 'DU/UD file-level pick flow (keep-content/keep-delete) stages correctly',
-                notebooks: ['02_base.ipynb', '02_current.ipynb', '02_incoming.ipynb'],
-                testModule: './duUdPickOneRegression.test.js',
-                requiresVSCode: true,
-            },
-            {
-                id: 'regression_au_ua_pick_one',
-                description: 'AU/UA add-only pick flow (apply-and-stage/cancel) stages correctly',
-                notebooks: ['02_base.ipynb', '02_current.ipynb', '02_incoming.ipynb'],
-                testModule: './auUaPickOneRegression.test.js',
+                testModule: './vscodeRegression.test.js',
                 requiresVSCode: true,
             },
         ],
